Check in Gemfile.lock
In order to do tests/deploys with the default Travis setup (that uses
`bundle install --deployment`), we need to check in a Gemfile.lock. This
is not normal done or advised in Gems that ship as a library, but makes
sense in this case.

Also put most of the development dependencies in the Gemfile instead of
the gemspec since that's a better place for them.