Compile and pack artifact when a Travis CI tag is detected
Previously, running `rake publish` on a tag would error with:

    /home/travis/.rvm/gems/ruby-1.9.3-p551/gems/dpl-1.7.11/lib/dpl/provider/s3.rb:37:in
    `chdir': No such file or directory - dist (Errno::ENOENT)

This is because when there's a tag, the `$TRAVIS_BRANCH` is the tag
name, not master, so compilation gets skipped. When compilation gets
skipped, there's no dist directory, and when there's no dist directory,
the upload fails.

This makes it so the compile still happens if there's a tag and
everything does what it did before if not.