Add vertical logo; Remove logo JS
* Add a vertical logo: ![Chef](https://www.chef.io/images/logo.svg)
* Remove the logo JavaScript and code for custom taglines.

The custom tagline stuff (where you could say `?site=Hey!` in the URL
and have it put the tagline underneath) was very cool but (1) used
JavaScript and (2) did not use Gotham for the font. (sorry @cnunciato!)

So, we lose that capability but also lose some complexity. The style
guide now shows you how to do the image tag with the SVG and the PNG
fallback with examples.