Publish merges to master and tagged commits
Currently, we build asset packages on tagged commits, but we exit before publishing the site. We should publish on tagged commits because the site displays the current version number and links to asset packages on GitHub releases. This makes that so.

It also fixes a bug preventing the site from being published on merges to master because of the TRAVIS_TAG environment variable containing an empty string (so always evaluating true).

Lastly it suppresses cloudfront activation during tests.