upgrade to Rails 4.2.10
Gems explicitly upgraded were:

* railties
* middleman
* rspec-rails
* git

This got us to Rails components at the latest point release of Rails
4.2.

The updated JSON library needed to be explicitly required by the Version
class file for the JSON constant to be available to it.

Set the middleman-cloudfront and middleman-s3_sync libraries to *not*
automatically be required by Bundler. These middleman plugins were used
for deploying the style site and assets to a CDN. We don't do that any
more, so there's no need to keep managing and updating these plugins and
the dependencies they bring with them.

Signed-off-by: Robb Kidd <robb@thekidds.org>